requirements-test.txt
# Test Dependencies for Veleron Whisper Voice-to-Text Project
# Python 3.13.7 compatible

# Core Testing Frameworks
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-timeout>=2.2.0
pytest-xdist>=3.5.0  # For parallel test execution

# Mocking and Fixtures
unittest-xml-reporting>=3.2.0

# Code Coverage
coverage>=7.4.0
coverage-badge>=1.1.0

# Testing Utilities
faker>=22.0.0  # For generating test data
hypothesis>=6.98.0  # Property-based testing

# Performance Testing
pytest-benchmark>=4.0.0

# Application Dependencies (needed for testing)
numpy>=1.26.0
torch>=2.0.0
openai-whisper>=20231117

# Audio Processing (for test fixtures)
sounddevice>=0.4.6
wave  # Built-in, but listed for clarity

# GUI Testing Dependencies
# Note: tkinter is built-in with Python

# System Integration Testing
pyautogui>=0.9.54
keyboard>=0.13.5
pystray>=0.19.5
Pillow>=10.0.0

# File I/O Testing
python-docx>=1.1.0  # If needed for future DOCX export testing

# Utilities
python-dateutil>=2.8.2
